feat(app): offload game moving parts
This module encapsulates SDL moving parts from main.
